+++ src/containers/DataTable/index.js
@@ -29,6 +29,7 @@ export function DataTable ({
 	buttonWithAuthz, sort, noItemsComponent,
 	customButton, customComponent,
 	customRowStyle, customRowClassName,
+	customCardBodyComponent,
 	limitValues = [10, 15, 25, 50]
 	}) {
 	const [filterValue, setFilterValue] = useState('');
@@ -160,7 +161,8 @@ export function DataTable ({
 					</CardHeader>
 
 					<CardBody className="data-table-card-body">
-						{ isLoading ? 
+						{customCardBodyComponent}
+						{isLoading ? 
 							<div style={{ margin: "2rem auto" }}><Spinner /></div> :
 							<Table
 								data={data.length > limit ? data.slice(0, limit) : data}